Exclude health check extension from code coverage
