+++ NotificationApi/NotificationApi/Health/HealthCheckExtensions.cs
@@ -1,9 +1,19 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Newtonsoft.Json;
 using NotificationApi.DAL;
 
 namespace NotificationApi.Health;
 
+[ExcludeFromCodeCoverage]
 public static class HealthCheckExtensions
 {
     public static IServiceCollection AddVhHealthChecks(this IServiceCollection services)
@@ -13,5 +23,43 @@ public static class HealthCheckExtensions
             .AddDbContextCheck<NotificationsApiDbContext>("Database VhNotificationsApi", tags: new[] {"startup", "readiness"});
             
         return services;
+    }
+    
+    public static IEndpointRouteBuilder AddVhHealthCheckRouteMaps(this IEndpointRouteBuilder endpoints)
+    {
+        endpoints.MapHealthChecks("/healthcheck/liveness", new HealthCheckOptions()
+        {
+            Predicate = check => check.Tags.Contains("self"),
+            ResponseWriter = HealthCheckResponseWriter
+        });
+
+        endpoints.MapHealthChecks("/healthcheck/startup", new HealthCheckOptions()
+        {
+            Predicate = check => check.Tags.Contains("startup"),
+            ResponseWriter = HealthCheckResponseWriter
+        });
+                
+        endpoints.MapHealthChecks("/healthcheck/readiness", new HealthCheckOptions()
+        {
+            Predicate = check => check.Tags.Contains("readiness"),
+            ResponseWriter = HealthCheckResponseWriter
+        });
+
+        return endpoints;
+    }
+
+    private static async Task HealthCheckResponseWriter(HttpContext context, HealthReport report)
+    {
+        var result = JsonConvert.SerializeObject(new
+        {
+            status = report.Status.ToString(),
+            details = report.Entries.Select(e => new
+            {
+                key = e.Key, value = Enum.GetName(typeof(HealthStatus), e.Value.Status),
+                error = e.Value.Exception?.Message
+            })
+        });
+        context.Response.ContentType = "application/json";
+        await context.Response.WriteAsync(result);
     }
 }
