+++ NotificationApi/NotificationApi/Startup.cs
@@ -1,22 +1,16 @@
 using System;
-using System.Linq;
-using System.Threading.Tasks;
 using FluentValidation;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
-using Newtonsoft.Json;
 using NotificationApi.Common.Configuration;
 using NotificationApi.Common.Util;
 using NotificationApi.DAL;
@@ -42,7 +36,7 @@ namespace NotificationApi
         {
             services.AddControllers()
                 .AddNewtonsoftJson();
-            services.AddSwagger();
+            services.AddVhSwagger();
             services.AddCors(options => options.AddPolicy("CorsPolicy",
                 builder =>
                 {
@@ -157,43 +151,12 @@ namespace NotificationApi
 
             app.UseEndpoints(endpoints =>
             {
-                endpoints.MapDefaultControllerRoute(); 
-                
-                endpoints.MapHealthChecks("/healthcheck/liveness", new HealthCheckOptions()
-                {
-                    Predicate = check => check.Tags.Contains("self"),
-                    ResponseWriter = HealthCheckResponseWriter
-                });
+                endpoints.MapDefaultControllerRoute();
 
-                endpoints.MapHealthChecks("/healthcheck/startup", new HealthCheckOptions()
-                {
-                    Predicate = check => check.Tags.Contains("startup"),
-                    ResponseWriter = HealthCheckResponseWriter
-                });
-                
-                endpoints.MapHealthChecks("/healthcheck/readiness", new HealthCheckOptions()
-                {
-                    Predicate = check => check.Tags.Contains("readiness"),
-                    ResponseWriter = HealthCheckResponseWriter
-                });
+                endpoints.AddVhHealthCheckRouteMaps();
             });
         }
         
-        private async Task HealthCheckResponseWriter(HttpContext context, HealthReport report)
-        {
-            var result = JsonConvert.SerializeObject(new
-            {
-                status = report.Status.ToString(),
-                details = report.Entries.Select(e => new
-                {
-                    key = e.Key, value = Enum.GetName(typeof(HealthStatus), e.Value.Status),
-                    error = e.Value.Exception?.Message
-                })
-            });
-            context.Response.ContentType = "application/json";
-            await context.Response.WriteAsync(result);
-        }
-
         private static void AddPolicies(AuthorizationOptions options)
         {
             options.DefaultPolicy = new AuthorizationPolicyBuilder()
